Update on-slot-render to TypeScript
